feat : Add faq component
