+++ src/components/Faq.js
@@ -1,0 +1,249 @@
+import React from 'react'
+import styled from "styled-components";
+
+const Container = styled.div`
+  border-top: #f7941e 5px solid;
+  width:100%;
+`;
+
+const Faq = styled.div`
+  width: 1132px;
+  margin-left: auto;
+  margin-right: auto;
+  margin-top: 64px;
+`;
+
+const Title = styled.div`
+  font-size: 30px;
+  letter-spacing: -1.5px;
+  float: left;
+  font-family: NanumSquare, serif;
+  font-weight: 800;
+  color: #175c8e;
+  height: 59px;
+  width: 100%;
+  text-align: left;
+`;
+
+const Desc = styled.div`
+  float: left;
+  text-align: left;
+  font-weight: 300;
+  letter-spacing: -1.2px;
+  line-height: 1.25;
+  font-size: 24px;
+  height: 80px;
+  margin-bottom: 12px;
+`;
+
+const AskButton = styled.button`
+  float: right;
+  display: block;
+  position: relative;
+  top: 15px;
+  margin-top: 10px;
+  background: #175c8e;
+  padding: 9px 15px;
+  letter-spacing: -0.6px;
+  color: white;
+  font-size: 12px;
+  text-decoration: none;
+`;
+
+const Tab = styled.div`
+  position: relative;
+  width: 1132px;
+  color: black;
+  overflow: hidden;
+  margin-bottom: 15px;
+  text-align: left;
+  display: table;
+  border-collapse: collapse;
+  
+  input {
+    height: 0;
+    display: none;
+    box-sizing: border-box;
+    padding: 0;
+    position: absolute;
+    z-index: -1;
+  }
+  
+  label {
+    position: relative;
+    display: block;
+    padding: 5px 5px 5px 26px;
+    line-height: 3;
+    cursor: pointer;
+    background: white;
+    font-size: 15px;
+    border: 1px solid #175c8e;
+    color: #252525;
+    letter-spacing: -0.8px;
+    margin-left: 1px;
+  }
+  
+  label::after {
+    position: absolute;
+    right: 0;
+    top: 0;
+    display: block;
+    width: 3em;
+    height: 3em;
+    line-height: 3;
+    text-align: center;
+    -webkit-transition: all .35s;
+    -o-transition: all .35s;
+    transition: all .35s;
+  }
+  
+  input:checked ~ .tab-content {
+     max-height: 10em;
+   }
+`;
+
+const UnderArrow = styled.img.attrs({
+  src: "https://static.koreatech.in/assets/img/bus_dropdown.png"
+})`
+  width: 15px;
+  float: right;
+  margin-right: 22px;
+  margin-top: 17px;
+`;
+
+const P = styled.p`
+  margin: 21px 26px 27px;
+  text-align: left;
+  float: left;
+  letter-spacing: -0.8px;
+`;
+
+const TabContent = styled.div`
+  max-height: 0;
+  overflow: hidden;
+  -webkit-transition: max-height .35s;
+  -o-transition: max-height .35s;
+  transition: max-height .35s;
+  color: #858585;
+  margin-left: 1px;
+  border: 1px solid #cbcbcb;
+  border-top: none;
+  margin-top: -1px;
+  width: 1132px;
+  background: #e4e4e4;
+`;
+
+const Answer = styled.b`
+  display: block;
+  float: left;
+  width: 17px;
+  height: 30px;
+`;
+
+const Pagination = styled.div`
+  margin-top: 35px;
+  margin-bottom: 40px;
+  background: #ffffff;
+  color: #858585;
+`;
+
+const ArrowButton = styled.button`
+  border-radius: 0;
+  padding: 7px 14px 7px 14px;
+  margin-left: 6px;
+  margin-right: 6px;
+  background: #ffffff;
+  border: 1px #edf0f3 solid;
+  color: #858585;
+  font-size: 13px;
+  letter-spacing: -0.7px;
+  cursor: pointer;
+  
+  &:hover {
+    background-color: #175c8e;
+    color: white;
+  }
+`;
+
+const Number = styled.button`
+  border-radius: 0;
+  padding: 7px 12px 7px 12px;
+  margin-left: 1px;
+  margin-right: 1px;
+  border: 1px #edf0f3 solid;
+  font-size: 13px;
+  cursor: pointer;
+  color: ${props => props.isNowPage ? 'white' : '#858585'};
+  background-color: ${props => props.isNowPage ? '#175c8e' : '#ffffff'};
+  
+  &:hover {
+    background-color: #175c8e;
+    color: white;
+  }
+`;
+
+export default function FaqPage(
+  {
+    faqList,
+    totalPageNum,
+    nowPageNum,
+    clickPrevButton,
+    clickNextButton,
+    clickPageNum
+  }) {
+  return (
+    <Container>
+      <Faq>
+        <Title>
+          FAQ
+        </Title>
+        <Desc>
+          아래 질문 외에 다른 궁금한 것이 있다면<br/>
+          직접 질문해주세요.
+        </Desc>
+        <AskButton onClick={() => {
+          window.location = "https://docs.google.com/forms/d/1pWoxCLm5YqTlcdiXThvyGlNu_7t5hojxgjH5FA8eypo/edit?usp=drive_web"
+        }}>
+          질문하기
+        </AskButton>
+        {faqList.map((faq, index) => {
+          return (
+            <Tab key={index}>
+              <input
+                type="radio"
+                name="tabs"
+                id={faq.id}/>
+              <label
+                htmlFor={faq.id}>
+                <span>
+                  <b>Q. </b>{faq.question}
+                </span>
+                <UnderArrow/>
+              </label>
+              <TabContent className="tab-content">
+                <P><Answer>A. </Answer>{faq.answer}</P>
+              </TabContent>
+            </Tab>
+          )
+        })}
+      </Faq>
+      <Pagination>
+        <ArrowButton onClick={clickPrevButton}>이전으로
+        </ArrowButton>
+        {[...Array(totalPageNum)].map((n, index) => {
+          return (
+            <span key={index}>
+              <Number
+                onClick={clickPageNum(index)}
+                isNowPage={(nowPageNum === index)}>
+                {index + 1}
+              </Number>
+            </span>
+          )
+        })}
+        <ArrowButton onClick={clickNextButton}>다음으로
+        </ArrowButton>
+      </Pagination>
+    </Container>
+  )
+}
