+++ src/modules/index.js
@@ -1,8 +1,10 @@
 import { combineReducers } from 'redux';
 import authReducer from './auth';
+import faqReducer from "./faq";
 
 const rootReducer = combineReducers({
   authReducer,
+  faqReducer
 });
 
 export default rootReducer;
